Time: 0 ms (100%), Space: 42.9 MB (69.15%) - LeetHub

+++ 0410-split-array-largest-sum/0410-split-array-largest-sum.java
@@ -2,15 +2,15 @@ class Solution {
     public int splitArray(int[] nums, int k) {
         if (k > nums.length) return -1;
 
-        long start = 0, end = 0;
+        int start = 0, end = 0;
 
-        for (long num : nums) {
+        for (int num : nums) {
             start = Math.max(start, num);
             end += num;
         }
 
         while (start <= end) {
-            long mid = start + (end - start) / 2;
+            int mid = start + (end - start) / 2;
 
             if (valid(nums, mid, k)) {
                 end = mid - 1;
@@ -19,10 +19,10 @@ class Solution {
             }
         }
 
-        return (int)start;
+        return start;
     }
 
-    private boolean valid (int[] nums, long ans, int k) {
+    private boolean valid (int[] nums, int ans, int k) {
         int pages = 0;
         int students = 1;
         for (int num : nums) {
